Maintain a unique ID throughout the lifetime of a surfel
The ID is stored in a previously unused field of the color/time texture.
When a surfel is created, its ID is set to the pixel location plus the
current timestamp multiplied by 307200. This guarantees ID uniqueness as
long as the 32-bit field does not overflow (allows at least 5000
frames).

This commit also fills in the Vertex structure (synonym of Surfel) with
the actual field names.